Dont fail when firewall-cmd is not available yet.
If the firewalld package is not installed yet, then the module fails
because firewalld tries to determine the state of the firewalld process
by using the firewall-cmd command, also the firewalld_zone resource type
will try and call the provider.exists? method in the generate method.

Both of these steps occur before the catalog is applied so before the
package resource can install the package.

This PR catches the exception when the command is missing and leaves
the @running instance variable set to nil.... When determining the
state of the firewalld service later in the Puppet run, the state
is re-checked if @running is set to nil.

Closes #96