Changed ruby require to Puppet::Type API references for types
Closes #93

As noted in #93, under certain circumstances, depending on the ordering
of which things are evaluated first in the manifests, Puppet 4.5.0+ can
throw an exception because it tries to re-register the types.  This was
discussed in https://tickets.puppetlabs.com/browse/PUP-6922

This commit removes the require statements for firewalld types and
instead references them using the Puppet::Type API to make sure they are
loaded.