Add defaults for services and ports
Having to specify service/port, zone and ensure for each service or port
is highly redundant and cumbersome. Make service and port default to the
resource name, ensure default to present and allow to set a default
target zone for services and ports via parameters to the firewalld
class. For ports the default protocol can also be set through a
firewalld class parameter.

Incidentally change the port spec test which seems to be incorrect
regarding parameter port accepting a hash.