Verify if firewalld is online before self.execute_firewall_cmd
self.instances method on the provider class did not verify if firewalld had a
correct state before issuing `firewall-cmd` commands.
This was a problem as we were trying to collect ipset informations
without checking if firewalld was ready.

So we move the `available?` implementation to the class method to be
able to verify the current firewalld state.